+++ di_lower_case.c
@@ -25,7 +25,7 @@ static void	di_print(t_specification spec, char *str)
 		ft_putchar(*str++);
 		g_return++;
 	}
-	while (spec.precision-- > ft_strlen(str))
+	while (spec.precision-- > (int)ft_strlen(str))
 	{
 		ft_putchar('0');
 		g_return++;
