+++ u_lower_case.c
@@ -15,7 +15,7 @@
 
 static void	u_print(t_specification spec, char *str)
 {
-	while (spec.precision-- > ft_strlen(str))
+	while (spec.precision-- > (int)ft_strlen(str))
 	{
 		ft_putchar('0');
 		g_return++;
